fix: Changed ID of Repository to Integer

+++ backend/src/main/java/fr/esipe/banxxy/repository/AccountRepository.java
@@ -3,5 +3,5 @@ package fr.esipe.banxxy.repository;
 import fr.esipe.banxxy.dao.AccountEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<AccountEntity, Long> {
+public interface AccountRepository extends CrudRepository<AccountEntity, Integer> {
 }
